+++ isy-aufrufkontext/src/test/java/de/bund/bva/isyfact/aufrufkontext/test/config/LeereTestConfig.java
@@ -1,0 +1,10 @@
+package de.bund.bva.isyfact.aufrufkontext.test.config;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+public class LeereTestConfig {
+
+}