+++ isy-ueberwachung/CHANGELOG.md
@@ -1,6 +1,10 @@
 # 2.2.0
 - `IFS-453`: Loglevel für isAlive-Datei-Ereignisse erhöht: INFO für Standardablageort wird genutzt; ERROR für Datei existiert bei Abfrage nicht.
 
+# 2.1.0
+- `IFS-436`: Entkopplung Health-Endpoint und hinzufügen Nachbarsystem-Check
+- `IFS-687`: Umstellung des Nachbarsystem-Checks auf RestTemplate statt WebClient
+
 # 2.0.0
 - `IFS-228`: Einführung von Spring Boot
 - `IFS-32`: Package-Name auf de.bund.bva.isyfact geändert